python 2 to 3

+++ utils.py
@@ -7,7 +7,7 @@ def merge(images, size):
 
     for idx, image in enumerate(images):
         i = idx % size[1]
-        j = idx / size[1]
+        j = int(idx / size[1])
         img[j*h:j*h+h, i*w:i*w+w] = image
 
     return img
